Utilize sweetAlert2 to handle errors and alerts
The sweetAlert2 CDN script was added to index.html. New functions errorMsg and Noresults were added to handle response errors and calls to the openBreweryDB which return no resluts, respectively. The getBreweries functions was updated to reflect the usage of these functions, as well as to account for the edge case error for calling for results for Alabama, which was causing a 'Cannot read property length of value null' error message for seemingly no absolute reason. Whatsoever.