Style and make responsive
Extensive work was done on the style sheet to bring color, positioning, animation and font to life. Among the changes include integration of the Cinzel Decorative and Mosterrat fonts from Google Fonts. Minor changes to the JS and html files were also made in order to aid with the styling changes.